Changed server version detection to use numbers
Changed the server version detection to use numbers rather than strings.
This allows for numerical comparisons to determine if a particular
function is supported.

The general convention for version numbering would just be the official
JC2-MP server version without the dot separators (e.g. Server version
`0.1.2` becomes `12.0`). The server version is represented as a
floating-point number to account for the possibility of additional
releases within the same version branch (e.g `0.1.2a`) should that ever
happen.

Server version `-1.0` represents server version is not yet known. This
can happen if a modified server is loaded or if a new server is loaded
with a older GUI. All functions are assumed to be compatible if the
server version is unknown.